+++ src/lustre/runtime/client/server_component.ffi.mjs
@@ -25,12 +25,13 @@ import {
 
 export class ServerComponent extends HTMLElement {
   static get observedAttributes() {
-    return ["route", "method"];
+    return ["route", "method", "csrf-token"];
   }
 
   #shadowRoot;
   #method = "ws";
   #route = null;
+  #csrfToken = null;
   #transport = null;
   #adoptedStyleNodes = [];
   #reconciler;
@@ -89,6 +90,9 @@ export class ServerComponent extends HTMLElement {
     switch (name) {
       case prev !== next && "route": {
         this.#route = new URL(next, location.href);
+        this.#csrfToken = this.#getCsrfToken();
+        this.#route.searchParams.set("csrf-token", this.#csrfToken);
+
         this.#connect();
         return;
       }
@@ -109,6 +113,22 @@ export class ServerComponent extends HTMLElement {
         }
 
         return;
+      }
+
+      case "csrf-token": {
+        if (prev !== next && this.#connected) {
+          this.#transport?.close();
+        }
+
+        this.#csrfToken = this.#getCsrfToken();
+
+        if (this.#route) {
+          this.#route.searchParams.set("csrf-token", this.#csrfToken);
+        }
+
+        if (this.#connected) {
+          this.#connect();
+        }
       }
     }
   }
@@ -302,6 +322,17 @@ export class ServerComponent extends HTMLElement {
     }
   }
 
+  #getCsrfToken() {
+    if (this.hasAttribute("csrf-token")) {
+      return this.getAttribute("csrf-token") || null;
+    } else {
+      const meta = document.querySelector('meta[name="csrf-token"]');
+      const token = meta.getAttribute("content");
+
+      return token || null;
+    }
+  }
+
   #connect() {
     if (!this.#route || !this.#method) return;
     if (this.#transport) this.#transport.close();
@@ -332,7 +363,12 @@ export class ServerComponent extends HTMLElement {
       );
     };
 
-    const options = { onConnect, onMessage, onClose };
+    const options = {
+      onConnect,
+      onMessage,
+      onClose,
+      csrfToken: this.#csrfToken,
+    };
 
     switch (this.#method) {
       case "ws":
@@ -635,6 +671,7 @@ class SseTransport {
 
 class PollingTransport {
   #url;
+  #csrfToken;
   #interval;
   #timer;
 
@@ -642,12 +679,14 @@ class PollingTransport {
   #onMessage;
   #onClose;
 
-  constructor(url, { onConnect, onMessage, onClose, ...opts }) {
+  constructor(url, { onConnect, onMessage, onClose, csrfToken, interval }) {
     this.#url = url;
+    this.#csrfToken = csrfToken;
+    this.#interval = interval ?? 5000;
+
     this.#onConnect = onConnect;
     this.#onMessage = onMessage;
     this.#onClose = onClose;
-    this.#interval = opts.interval ?? 5000;
 
     this.#fetch().finally(() => {
       this.#onConnect();
@@ -663,7 +702,15 @@ class PollingTransport {
   }
 
   #fetch() {
-    return fetch(this.#url)
+    const headers = Object.assign(
+      {},
+      // If a CSRF token is provided, include it as a request header as checking
+      // headers is more common than query param for CSRF protection in traditional
+      // HTTP requests.
+      this.#csrfToken && { "x-csrf-token": this.#csrfToken },
+    );
+
+    return fetch(this.#url, { headers })
       .then((response) => response.json())
       .then(this.#onMessage)
       .catch(console.error);